+++ examples.py
@@ -37,18 +37,18 @@ api.Boards.change_work_item_state(testcase_id, WorkItemsStatesDef.TestCase.Close
 
 # Repo Examples:
 ## Clone repository without history
-api.Repo.clone_repository(tempfile.gettempdir(), branch="main", depth=1, submodules=True)
+api.Repos.clone_repository(tempfile.gettempdir(), branch="main", depth=1, submodules=True)
 
 ## Get list of all available branches
-all_branches_list = api.Repo.get_all_branches()
+all_branches_list = api.Repos.get_all_branches()
 
 ## Check available pull requests
-prs = api.Repo.get_active_pull_requests()
+prs = api.Repos.get_active_pull_requests()
 
 ## Create Pull Request
-pr_id = api.Repo.create_pr("Test PullRequest", "TestBranch", "main", "Testing API Request.")
+pr_id = api.Repos.create_pr("Test PullRequest", "TestBranch", "main", "Testing API Request.")
 ## Add Reviewer to PR
-api.Repo.add_pr_reviewer(pr_id, "user1@gmail.com")
+api.Repos.add_pr_reviewer(pr_id, "user1@gmail.com")
 
 # Agents Examples:
 ## Add Agents userCapabilies