+++ CHANGELOG.md
@@ -8,9 +8,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Docstrings and full test coverage for AzApi Repos
 - Batch script to run tests and generate coverage logs
+- Exception raising when uninitiated component is used.
 
 ### Changed
 - Added handling of downloading repository from custom url by keyword arg in `clone_repository` method.
+- Changed Repo to Repos name component to match project convention.
+
+### Fixed
+- Fixed hardcoded email in AzApi class
+
 
 ## [0.1.1] - 
 ### Added